// src/interfaces/ubi/community.ts
export interface CommunityAttributes {
    id: number; // Note that the `null assertion` `!` is required in strict mode.
    publicId: string; // TODO: to be removed
    requestByAddress: string;
    contractAddress: string | null;
    name: string;
    description: string;
    descriptionEn: string | null;
    language: string;
    currency: string;
    city: string;
    country: string;
    gps: {
        latitude: number;
        longitude: number;
    };
    email: string;
    visibility: 'public' | 'private';
    coverImage: string; // TODO: to be removed
    coverMediaId: number;
    status: 'pending' | 'valid' | 'removed'; // pending / valid / removed
    review: 'pending' | 'in-progress' | 'halted' | 'closed';
    started: Date; // TODO: to be removed
    proposalId: number | null;

    // timestamps
    createdAt: Date;
    updatedAt: Date;
    deletedAt: Date | null;

    metrics?: UbiCommunityDailyMetrics[]; // TODO: to be removed
    cover?: AppMediaContent;
    proposal?: AppProposal;
    contract?: UbiCommunityContract; // TODO: to be removed
    state?: UbiCommunityState; // TODO: to be removed
    storyCommunity?: StoryCommunity[]; // TODO: to be removed
    suspect?: UbiCommunitySuspect[]; // TODO: to be removed
    beneficiaries?: BeneficiaryAttributes[]; // TODO: to be removed
    // promoter?: UbiPromoter;
    claimLocation?: UbiClaimLocation[]; // TODO: to be removed
    demographics?: UbiCommunityDemographics[]; // TODO: to be removed
    dailyState?: UbiCommunityDailyState[]; // TODO: to be removed
}